frontend/requirements.txt
streamlit
requests
fpdf
Pillow
